+++ src/lib.rs
@@ -228,9 +228,7 @@
 //! [stderr]: https://doc.rust-lang.org/std/io/fn.stderr.html
 //! [flush]: https://doc.rust-lang.org/std/io/trait.Write.html#tymethod.flush
 
-#[cfg(windows)]
-pub use utils::functions::supports_ansi;
-pub use utils::{Command, ErrorKind, ExecutableCommand, Output, QueueableCommand, Result};
+pub use utils::{Command, ErrorKind, ExecutableCommand, QueueableCommand, Result};
 
 /// A module to work with the terminal cursor
 pub mod cursor;