+++ src/utils/sys.rs
@@ -1,5 +1,5 @@
 #[cfg(windows)]
-pub mod winapi;
+pub(crate) mod windows;
 
 #[cfg(unix)]
-pub mod unix;
+pub(crate) mod unix;